+++ src/app/(with-nav)/(with-header)/character/page.tsx
@@ -1,16 +1,16 @@
 'use client';
 
-import { usePostExperienceMutation } from '@/hooks/mutations/use-post-experience-mutation';
+import { usePostExperienceMutation } from '@/features/character/hooks/use-post-experience-mutation';
 import { useCharacterStore } from '@/store/use-character-store';
 
 const CharacterPage = () => {
   const characterId = useCharacterStore((state) => state.characterId);
 
-  const { mutate: updateExp } = usePostExperienceMutation();
+  const { mutate: expUpdateMutation } = usePostExperienceMutation();
 
   const handleExperienceUp = () => {
     if (characterId === null) return; 
-    updateExp({ characterId, amount: 10 });
+    expUpdateMutation({ characterId, amount: 10, history: 'history' });
   };
 
   return (
